+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: `EditorialCard`: Add component.
 - Feature: `LongArrowIconNext`: Add component.
+- Feature: `ProjectCard`: Add `overlayText` prop with empty default.
+- Feature: `ProjectCard`: Add `hoverableTitle` prop, defaults to false.
 - Feature: Add the following components:
   - `ChatLayout`
   - `Discussion`